feat: Implement TailwindCSS
This adds TailwindCSS (www.tailwindcss.com) functionality to the page,
obviating the need for using inline styles and letting us rely on
utility classes instead. Because I've somewhat tried to maintain the
look of the original template, I've created some awkward nonce-classes
that would not be part of any real deployment.